backend/product/requirements.txt
asgiref==3.7.2
celery==5.3.6
click==8.1.7
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
h11==0.14.0
packaging==23.2
pika==1.3.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
uvicorn==0.24.0.post1
